Move API Discord guild check from controller to middleware
